[refactor] Sections 일급 컬렉션 적용
